backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
openai==1.12.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
stripe==10.0.0

